query-editor: use entries on the search bar
It's important to give feedback even when the search popover is closed
about which filters are in place. This is only achievable if the search
bar shows some labels tags as gnome-photos or gnome-documents does.

Unfortunately we don't have this tool in gtk+ yet, so we need to use
libgd.

So implement the query-editor with a custom GdtaggedEntry and update
the libgd subrepository to apply the latest style changes.